+++ CHANGELOG.md
@@ -8,6 +8,21 @@
 
 ### Changed
 
+### Changed
+
+  - Edge detail - Do not propose execution-mode in add or revoke SSH credentials modal
+
+## [2.25.0] - 2021-10-21
+
+### Added
+
+  - Support for OpenStack infrastructures and credentials
+
+### Changed
+
+  - Fix selection of default NuvlaBox version during creation
+  - Do not display hashed password type credentials in credentials page 
+
 ## [2.24.0] - 2021-10-13
 
 ### Changed